refactor(ui): move table sort primitives into their own module
VideoListSortHeader.tsx exported the SortableHeader component next to the
column contract, the persisted-value guard and the comparator, so Vite Fast
Refresh dropped the header's state on every edit (4 of the 6
react-refresh/only-export-components warnings).

TableSortKey/SortDirection/TableSort, DEFAULT_TABLE_SORT, NATURAL_DIRECTION,
isTableSort and sortValue now live in videoListSort.ts. Pure move, no
behaviour change; VideoListTable imports the primitives from there.
