+++ src/shared/components/ui/VideoListSortHeader.tsx
@@ -1,65 +1,13 @@
 import React from "react";
 import { ArrowDown, ArrowUp, ArrowUpDown } from "lucide-react";
 import { useTranslation } from "react-i18next";
-import type { VideoData } from "@/src/features/videos";
+import type { TableSort, TableSortKey } from "./videoListSort";
 
 /**
- * Sorting primitives for the analyser result table: the column contract, the
- * numeric comparator and the clickable header cell. Kept beside the table so
- * VideoListTable stays within the file-size guideline.
+ * The clickable header cell of the analyser result table. Kept beside the table
+ * so VideoListTable stays within the file-size guideline; the column contract
+ * and comparator it works with live in `videoListSort.ts`.
  */
-
-export type TableSortKey = "rank" | "upload" | "views" | "velocity" | "engagement" | "score";
-export type SortDirection = "asc" | "desc";
-
-export interface TableSort {
-  key: TableSortKey;
-  dir: SortDirection;
-}
-
-/** Rank ascending reproduces the order the analyser handed the rows over in. */
-export const DEFAULT_TABLE_SORT: TableSort = { key: "rank", dir: "asc" };
-
-/** Direction a column starts in on its first click — biggest/newest first. */
-export const NATURAL_DIRECTION: Record<TableSortKey, SortDirection> = {
-  rank: "asc",
-  upload: "desc",
-  views: "desc",
-  velocity: "desc",
-  engagement: "desc",
-  score: "desc",
-};
-
-/** Guard for the persisted value — storage is untrusted input. */
-export function isTableSort(value: unknown): value is TableSort {
-  if (!value || typeof value !== "object") return false;
-  const v = value as Record<string, unknown>;
-  return (
-    typeof v.key === "string" && v.key in NATURAL_DIRECTION && (v.dir === "asc" || v.dir === "desc")
-  );
-}
-
-/**
- * Numeric sort key for a column. Missing optional metrics sort below zero so
- * they end up last on a descending sort instead of tying with a real 0.
- */
-export function sortValue(video: VideoData, rank: number, key: TableSortKey): number {
-  switch (key) {
-    case "upload":
-      return video.publishedTimestamp;
-    case "views":
-      return video.views;
-    case "velocity":
-      return Number.isFinite(Number(video.viewsPerHour)) ? Number(video.viewsPerHour) : -1;
-    case "engagement":
-      return video.engagementRate ?? -1;
-    case "score":
-      return video.trendingScore;
-    case "rank":
-    default:
-      return rank;
-  }
-}
 
 const ALIGN_CLASS: Record<"left" | "center" | "right", string> = {
   left: "justify-start",
